Remove dead code, Improve consistency and minor issues
This changeset tightens cache file path handling in phlib by validating leaf names, honoring native vs Win32 path modes consistently, and using the correct directory creation path. It also fixes several functional regressions across the app and plugins: module list filter refresh after highlight toggles, ETW endpoint/port mapping, aggregate sort field usage, graph child window clipping, safer uninstall status text handling, previous-instance shutdown in setup, and improved window-message timeout handling. Additional updates clean up initialization/casts, remove dead code, and address small correctness issues in formatting and utility logic.

+++ phlib/native.c
@@ -227,7 +227,7 @@ NTSTATUS PhMergeSystemAcls(
 
     // Allocate new ACL (header + aligned payload).
 
-    requiredSize = sizeof(ACL) + ALIGN_UP_BY(requiredSize, sizeof(ULONG));
+    requiredSize = (ULONG)sizeof(ACL) + ALIGN_UP_BY(requiredSize, sizeof(ULONG));
 
     if (requiredSize > USHORT_MAX)
         return STATUS_INVALID_PARAMETER;
